+++ Saves/NonSavedData.cs
@@ -1,0 +1,6 @@
+﻿namespace WECCL.Saves;
+
+public class NonSavedData
+{
+    public static List<Character> DeletedCharacters { get; set; } = new();
+}