+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## GV 1.60:
 
+### Release 1.7.0
+- Added support for adding and removing characters from the search screen.
+- Deleted characters are exported to the `Purgatory` folder after saving.
+- Separated the modded save file from the vanilla save file (can be changed in the config).
+
 ### Release 1.6.2
 - Fixed mappings for custom costumes. The system has been reworked to automatically detect future changes to the game's
   vanilla costume count.